refactor: split wifi_setup.py (39KB) into 3 focused modules
- wifi.py: Shared WiFi utilities (scan, connect, hotspot, hostname)
  with wifi_interface parameter for platform abstraction
- status_server.py: CameraStatusServer + StatusHandler (post-setup)
  with injectable wifi_interface and thermal_path
- wifi_setup.py: WifiSetupServer + SetupHandler (first-boot only),
  slimmed from 1002 lines to ~280 lines

Backward-compatible re-exports in wifi_setup.py so existing
imports continue to work. Tests updated to use wifi module
functions directly.

199 tests pass, 67% coverage.
